requirements.txt
Django==5.2.8
python-dotenv==1.2.1
requests>=2.28
bootstrap4
